Rename eval components to evaluators
